Have pre-commit hook ignore comments in .gitignore
While trying to add my own customization to the .gitignore, I had an
issue where the pre-commit hook was seeing a reference to an issue in
another project inside a comment as a directory and producing an error.
This fixes my issue by ignore lines that begin with # when parsing the
.gitignore. Note that this will still pick up inline comments, so there
should probably be a better fix than this.